added update and delete functions but no yet working comment and like

+++ src/app/data/back-end.service.ts
@@ -16,43 +16,57 @@ export class BackEndService {
 
   //saving data from creating post
   saveData(post: Post){
-    this.http.post(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts/${post.postId}.json`, post)
+    this.http.post<{name: string}>(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts.json`, post)
     .subscribe((res)=>{
       console.log(res);
-      this.fetchData().subscribe();
+      post.postId = res.name;
+      this.postService.addPost(post);
     })
   }
 
   //fetching data from firebase for viewing in home
   fetchData(){
-    return this.http.get('https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts.json')
-    .pipe(map(responseData => Object.values(responseData)),
-          tap((listsOfPosts: Post[]) => {
-            this.postService.setPosts(listsOfPosts);
-        })
+    return this.http.get<Record<string, Post>>('https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts.json')
+    .pipe(map(responseData => {
+      const postsArray: Post[] = [];
+      for (const key in responseData) {
+        if (responseData.hasOwnProperty(key)) {
+          let newPost = new Post(responseData[key].title, responseData[key].imgPath, responseData[key].description, responseData[key].dateCreated, responseData[key].numberOfLikes, responseData[key].comments, responseData[key].ownerId, responseData[key].userEmail);
+          newPost.postId = key;
+          postsArray.push(newPost);
+        }
+      }
+      return postsArray;
+    }),
+    tap((listsOfPosts: Post[]) => {
+      this.postService.setPosts(listsOfPosts);
+    })
     )
   }
  
   //updating the data 
   updateData(postId: string, updatedPost: Post) {
-  this.http.put(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts/${postId}.json`, updatedPost)
-    .subscribe(response => {
-        console.log(response);
-    });
-}
+    this.http.put(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts/${postId}.json`, updatedPost)
+      .subscribe(response => {
+          console.log(response);
+          this.postService.updatePost(parseInt(postId), updatedPost);
+      });
+  }
 
 
   //deleting data from firebase
   deleteData(postId: string) {
-    this.postService.deleteButton(postId);
-    this.http.delete(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts/${postId}.json`)
-        .subscribe(response => {
-            console.log(response);
-            this.fetchData().subscribe((posts) => { 
-                this.postService.setPosts(posts);
-            });
-        });
-}
+    if (window.confirm('Are you sure you want to delete this post?')) {
+      this.postService.deleteButton(postId);
+      this.http.delete(`https://crud-app-f0d6e-default-rtdb.asia-southeast1.firebasedatabase.app/posts/${postId}.json`)
+          .subscribe(response => {
+              console.log(response);
+              this.fetchData().subscribe((posts) => { 
+                  this.postService.setPosts(posts);
+              });
+          });
+    }
+  }
 
 
   //comment push to firebase
